refactor(web): split HubDashboard.tsx — extract module configs, bento card, dashboard cards
